# api/openapi.yaml
openapi: 3.0.3
info:
  title: AxonOps Schema Registry
  description: |
    A high-performance Kafka Schema Registry compatible with Confluent Schema Registry API.
    Supports Avro, Protobuf, and JSON Schema formats.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8081
    description: Local development server
...
paths:
...
  /me:
    get:
      summary: Get current user info
      operationId: getCurrentUser
      tags:
        - Account
      security:
        - basicAuth: []
        - apiKey: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/password:
    post:
      summary: Change own password
      description: Change the password for the currently authenticated user. Requires the current password for verification.
      operationId: changePassword
      tags:
        - Account
      security:
        - basicAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed successfully
        '400':
          description: Invalid request (e.g., password too weak)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Admin endpoints - User Management
...
  /admin/users/{id}:
    get:
      summary: Get user by ID
      operationId: getUser
      tags:
        - Admin
      security:
        - basicAuth: []
        - apiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update user
      operationId: updateUser
      tags:
        - Admin
      security:
        - basicAuth: []
        - apiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete user
      operationId: deleteUser
      tags:
        - Admin
      security:
        - basicAuth: []
        - apiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: User deleted
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Admin endpoints - API Key Management
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: integer
          description: Error code
        message:
          type: string
          description: Error message

    # Admin schemas - Users
...
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          description: New password
        role:
          type: string
          enum: [super_admin, admin, developer, readonly]
          description: User role
        enabled:
          type: boolean
          description: Whether the user is enabled
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          description: Email address
        role:
          type: string
          description: User role
        enabled:
          type: boolean
          description: Whether the user is enabled
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
...
    ChangePasswordRequest:
      type: object
      required:
        - old_password
        - new_password
      properties:
        old_password:
          type: string
          format: password
          description: Current password for verification
        new_password:
          type: string
          format: password
          description: New password to set

    # Admin schemas - API Keys
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        HTTP Basic Authentication. Supports two modes:
        1. Username/password: Standard basic auth with username and password
        2. API key via Basic Auth: Use API key as username with empty password
           (compatible with Kafka producers/consumers)
           Example: Authorization: Basic <base64(api_key:)>
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed via X-API-Key header
security:
  - basicAuth: []
  - apiKey: []